set up routing and reorg
