+++ .env
@@ -1,0 +1,1 @@
+REACT_APP_NOT_SECRET_CODE=https://edvn99kbqa.execute-api.us-west-2.amazonaws.com/dev
