+++ src/App.js
@@ -1,0 +1,13 @@
+
+import Main from './Main';
+
+
+function App() {
+    return (
+        <div className="App">
+            <Main />
+        </div>
+    );
+}
+
+export default App;
