--- src/Main.test.js
@@ -1,8 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import Main from './Main';
-
-test('renders learn react link', () => {
-  render(<Main />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
