+++ src/pages/Home/Title.css
@@ -1,7 +1,3 @@
-a {
-  color: var(--theme-contrast);
-}
-
 header {
   line-height: 1.2;
   position: absolute;
@@ -61,6 +57,7 @@ header {
   transition: none 200ms ease-out;
   transition-property: color, background;
   pointer-events: auto;
+  background-color: unset;
 }
 .button-black {
   color: var(--theme-dark-text);
