+++ src/pages/TIForm/TIForm.css
@@ -1,0 +1,7 @@
+.ti-form {
+    padding: 5px 20px;
+}
+
+.ti-form fieldset {
+    margin: 20px 0;
+}
